// src/app/shared/model/usuario-model.ts
export interface IUsuario {
    id: number;
    tipoDocumento: string;
    identidad: number;
    nombre: string;
    apellido: string;
    email: string;
    genero: string;
    estadoCivil: string;
    telefono: number;
    movil: number;
    direccion: string;
    tarjetaProfesional: string;
    especialidad: string;
    ciudad: number;
    departamento: number;
    fechaNacimiento: string;
    estado: string;
    fechaRegistro: string;
    usuarioRegistro: string;
}